system-helper: Add CancelPull D-Bus method
If there is a pull failure in a directory created by CreatePullCacheDir
D-Bus method of system-helper, there is no way to go back to the system
helper and notify it to cleanup. Therefore, CancelPull is required
on the pull failure error path, so that the bindfs and friends
can be cleaned up nicely and prevent any dangling mounts.

https://phabricator.endlessm.com/T15371