Integrate --system pull with revokefs filesystem.
system-helper: Add a new helper method RevokefsFd which stands responsible
for spawning the backend part of the revokefs filesystem. It takes care
of creating a cache location for the backing directory and creating a mount
point where the client can write to. The other end socket is transferred over
D-Bus to the client which spawns the fuse filesystem in the mountpoint and
starts the pull.

The client upon receiving the socket spawns the fuse filesystem and create
a child repo for the pull. In any case of failure, it fallbacks on the current
code path (which causes duplication).

The backing dir is owned by the "flatpak" user during the pull.
It's ownership is then canoncalized with owner=root and permissions
as per bare-user-only. Now we have fulfilled all the criteria to
hardlink the repo. See [1].

[1] https://github.com/ostreedev/ostree/pull/1776