system-helper: Add CancelPull helper method
If there is a pull failure in a child repo created on revokefs mount,
there is no way to go back to the system helper and notify it to cleanup.
Therefore, CancelPull is required on the pull failure error path, so
that the ongoing pull can be cleaned up nicely and prevent any dangling
mounts and subprocesses.