installation: Return main app as updatable if related extensions are missing
While updating, if the related extension is missing on
the installation and the main app is installed, FlatpakTransaction
tends to "repair" the main app by automatically downloading
the related extension again and restoring the overall functionality
of the app.

The related extension concerned that are the ones associated with
"should-download" to TRUE only.

Hence, teach the libflatpak API to do that same, so that clients
like gnome-software can mark those apps as updatable, if their
related extensions is missing.